concept-explorer/requirements.txt
streamlit>=1.30
pyvis>=0.3.2
pymupdf>=1.23.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0
mistralai>=1.0.0
networkx>=3.2
